third-commit

+++ FrontEnd/modal.js
@@ -1,17 +1,19 @@
+//call 1 time the function 
 
-let jsonDataCatModal;
-let recallWorksVar;
+
 document.getElementById("area-admin").addEventListener('click', function () {
     var openModalBtn = document.getElementById('my-modal')
     openModalBtn.style.display = 'block';
-    if (openModalBtn.style.display === 'block') {
-        let recallWorksVar = async function recallWorks() { 
+    if (openModalBtn.style.display === 'block' ) {
+      async function recallWorks() { 
             const responseModal = await fetch('http://localhost:5678/api/works/')
             const jsonDataCatModal = await responseModal.json();
-            jsonDataCatModal.forEach(arrayJsonMod => {
+console.log(jsonDataCatModal)
+            jsonDataCatModal.forEach(arrayJsonMod  => {
                 const showAllPhoto = document.getElementById('gallery-edit')
                 const imgFigureMod = document.createElement('figure')
                 imgFigureMod.classList.add("figure-modal")
+        
                 const trashBtn = document.createElement('div')
                 trashBtn.innerHTML = `<i class="fa-solid fa-trash-can"></i>`
                 trashBtn.classList.add("trash-btn")
@@ -30,40 +32,57 @@ document.getElementById("area-admin").addEventListener('click', function () {
                 imgFigureMod.appendChild(arrowBtn)
             })
             document.querySelectorAll(".trash-btn").forEach(trash => {
-                trash.addEventListener("click", function () {
-                    remuoveEleFromDom()
+                trash.addEventListener("click", async  function () {
+                    const imgFigureModCall = document.querySelector("figure")
+                    imgFigureModCall.setAttribute("data-id", 1)
+                    if(responseModal.ok){
+
+                        let token = sessionStorage.getItem('reponseLogin')
+                        token = JSON.parse(token)
+                        console.log(token.userId)
+                        let reponseForCancell = await fetch('http://localhost:5678/api/works/${Id}', {
+                            method: 'delete',
+                            headers: {
+                                Authorization: `accept: ${token}`,
+                                'Content-Type': "application/json"
+                            },
+                            body: JSON.parse(token.userId)
+                        })
+
+                     imgFigureModCall.remove()
+                    }
+                    console.log(jsonDataCatModal)
+                 
                 })
             })
-   
+          /*   
+            async function remuoveEleFromDom() {
+                let token = sessionStorage.getItem('reponseLogin')
+                token = JSON.parse(token)
+                console.log(token.userId)
+                let reponseForCancell = await fetch('http://localhost:5678/api/works/${Id}', {
+                    method: 'delete',
+                    headers: {
+                        Authorization: `accept: ${token}`,
+                        'Content-Type': "application/json"
+                    },
+                    body: JSON.parse(token.userId)
+                })
+              let workEle = target.closest('.figure-modal')
+                if (token.userId === 1 && reponseForCancell.status === 400) {
+                workEle.remove()
+                    
+                    
+                }
+            } */
+
+
         }
         recallWorks()
     }
 })
 
-async function remuoveEleFromDom() {
-    let token = sessionStorage.getItem('reponseLogin')
-    token = JSON.parse(token)
-    console.log(token.userId)
-    let reponseForCancell = await fetch('http://localhost:5678/api/works/${Id}', {
-        method: 'delete',
-        headers: {
-            Authorization: `accept: ${token}`,
-            'Content-Type': "application/json"
-        },
-        body: JSON.parse(token.userId)
-    })
-    const eventRemuve = jsonDataCatModal;
-    console.log(eventRemuve)
-    let parsedReponseCancell = JSON.stringify(reponseForCancell)
-    console.log(reponseForCancell)
-    console.dir(parsedReponseCancell)
-    if (token.userId === 1 && reponseForCancell.status === 400) {
-        const showAllPhoto = document.getElementById('gallery-edit')
-        showAllPhoto.innerHTML = ""
-        
-        
-    }
-}
+
 
 
 document.getElementById('close-btn').addEventListener('click', function () {
